requirements.txt
tensorflow
Pillow
tkinter
numpy
matplotlib
opencv-python